% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglefix.R
\name{anglefix}
\alias{anglefix}
\title{Angle difference fixer}
\usage{
anglefix(a, deg = F)
}
\arguments{
\item{a}{an angle in rad from -2pi (anticlockwise) to +2pi (clockwise)}

\item{deg}{boolean if you want to enter it in degrees. I don't think it works yet}
}
\value{
An angle that is always between -pi and pi
}
\description{
Resolves discontinuity artefacts when calculating angle differences by always selecting
the shortest of the two rotations.
}
\details{
Takes an angle (in rad: from -2pi to 2pi; or in degrees: from -360 to 360) and if it's larger than
half a circle replaces it with the shortest distance between the same points (in the opposite direction).

I don't think the degree version quite works yet.
}

