% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figuredf.R
\name{figuredf}
\alias{figuredf}
\title{Figure reshaping for plotting purposes}
\usage{
figuredf(fig)
}
\arguments{
\item{figures}{a figure list, output of \code{flyfigures}}

\item{trx}{a track data frame}
}
\value{
a data frame including \code{x} and \code{y} values stacked by \code{start} (figure start time: constant for each figure
and consistent with experimental frame) and \code{step} (counter from first to last frame of the figure; \code{start} + \code{step}
should reconstruct data point frame), and labelled with \code{id}.
}
\description{
Generates a plot-friendly data frame from the list output of \code{flyfigures}.
}
\details{
Requires the list format to map each figure to a specific fly (\code{id}) and moment (\code{starttime}).
If the optional argument \code{trx} is also passed, all the other attributes attached to trx are also
included by merging.
}

