% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filldiff.R
\name{filldiff}
\alias{filldiff}
\title{NA-filled diff}
\usage{
filldiff(x, gap = 1)
}
\arguments{
\item{x}{a vector of any type \code{diff} is happy with}

\item{gap}{the gap to be passed to \code{diff}}
}
\value{
A difference vector with \code{NA}s introduced to make it up to the same size as the input
}
\description{
Applies the diff function but pads it with \code{NA}s to retain original length.
}
\details{
When the gap is an odd number the extra \code{NA} goes at the start.
}

