% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyangle.R
\name{flyangle}
\alias{flyangle}
\title{Track angle calculator}
\usage{
flyangle(trx, framelag = 6)
}
\arguments{
\item{trx}{a track data frame (only requires CTrax output variables)}

\item{framelag}{the number of frames to derive velocity across; defaults to 6}
}
\value{
A numeric vector matching the structure of \code{trx} containing the angle value for each object and timepoint, or \code{NA}s when it can't be calculated.
}
\description{
Calculates the angle of the velocity vector in rad by comparing centroid location across \code{framelag}.
}
\details{
For each timepoint t the velocity vector is calculated between \code{t-framelag/2} and \code{t+framelag+2}.
Currently this leaves \code{framelag/2 NA}s at the beginning and end of each track. This causes all sorts of problems
and may be best addressed somehow in the future.

The angle is expressed in reference to the (arbitrary) coordinate system of the video and therefore this value is not overly useful.
}

