% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flychain.R
\name{flychain}
\alias{flychain}
\title{Transition matrix generation from event series}
\usage{
flychain(events, chainlag)
}
\arguments{
\item{events}{a vector of consecutive events (for example, the \code{whatis} column of a \code{flyclass}-processed track data frame)}

\item{chainlag}{frame lag across which to compare}
}
\value{
A transition matrix, normalised to have rows summing to 1 for further probability work.
}
\description{
Takes a vector of consecutive events and compares each to the event \code{chainlag} frames away.
}
\details{
This is a bit of a wrapper function because the format of \code{table} is really awkward and I expect I would forget
how to handle it and have to re-google it avey time.
}

