% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyclass.R
\name{flyclass}
\alias{flyclass}
\title{Track event classifier}
\usage{
flyclass(trx, sep = exp(-1), edge = 12, morph = 7, runlag = 6)
}
\arguments{
\item{trx}{a track data frame}

\item{sep}{the discriminant speed value between runs and stands. Set at e^-1 on the basis of eyeballed data.}

\item{edge}{the number of pixels from the rim of the arena to be categorised as edge.}

\item{morph}{the window for morphological operations}

\item{runlag}{the \code{framelag} passed to \code{flyrun} in case the speed column is missing}
}
\value{
a track data frame that includes the boolean columns \code{isedge} and \code{isrun} and the factor column \code{whatis}
}
\description{
Uses speed and location information to classify entries into \code{stand}, \code{run} or \code{edge} events.
}
\details{
The classification is based on the columns \code{speed} (output of \code{flymoves} or \code{safemoves}) and \code{radius} (output of \code{trxgarnish}).
If these are not already included, they will be calculated ad hoc. Rows with NAs (with the exception of \code{turn}, which is tolerated as it is not used for calculations but has more NAs than anything else)
are removed, then each point is categorised as either edge or not (by a threshold pixel distance \code{edge})
and either run or not (by a threshold pixel/frame speed \code{sep}). Short running/standing bouts are eliminated
by morphological opening and closing over a window of \code{morph} frames. Then, each entry is categorised as either
\code{edge}, \code{run} or \code{stand} in the classifier variable \code{whatis} - note that there is currently no distinction between
runs and stands in the edge region.

Note: currently the opening/closing is not run by id, so there is a bit of spillage
between consecutive tracks, but it should only affect ~4 frames on each side. It would still be best to run the function via ave.
}

