% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyfigures.R
\name{flyfigures}
\alias{flyfigures}
\title{Track segmentation and realignment}
\usage{
flyfigures(trx, stepsize = 6, output = "list")
}
\arguments{
\item{trx}{a track data frame}

\item{stepsize}{the size of the time window (in frames)}

\item{output}{whether to output the minimal data (a matrix containing the x,y data as the real and imaginary
parts of a complex number) or a more intuitive but somwehat fiddlier list format.}
}
\value{
either a \code{frame} by \code{stepsize} complex matrix (if \code{output='complex'}) where Re is
x and Im is y, or a list containing a matrix for \code{x}, one for \code{y}, and vectors for \code{id} and
\code{starttime} to match each matrix row to a frame and fly unambiguosly.
}
\description{
Aligns \code{stepsize} consecutive frames from each start to a reference coordinate system where start is
always (0,0) and end is always (x,0).
}
\details{
From each frame, \code{stepsize} frames are selected and the position data (\code{x} and \code{y})
are aligned (shifted and rotated) to a common reference system, yielding transformed \code{x} and \code{y}.
These are collected into a matrix of \code{frame} by \code{stepsize} elements (thus oversampling the data).
}

