% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flymoves.R
\name{flymoves}
\alias{flymoves}
\title{Track motion parameter calculator}
\usage{
flymoves(trx, speedlag = 6, anglelag = 6, turnlag = 6)
}
\arguments{
\item{trx}{a track data frame (CTrax values)}

\item{speedlag}{frame lag for the calculation of speed}

\item{anglelag}{frame lag for the calculation of the velocity vector angle}

\item{turnlag}{frame lag for the calculation of the angular velocity}
}
\value{
A track data frame including the columns \code{speed}, \code{angle} and \code{turn}. These will contain \code{NA} chunks (longer for \code{angle}).
}
\description{
A wrapper function that calls \code{\link{flyrun}}, \code{\link{flyangle}} and \code{\link{flyturn}} and appends them to a track data frame.
}
\details{
Each motion function can receive a specific \code{framelag} value; by default they are all at 6.
This function is kept separate from \code{\link{trxgarnish}} so that it can be run with different \code{framelag} parameters without recalculating all the location variables as well.
In the future \code{angle} could be optional as it's a bit of a lame duck variable.
}

