% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyspin.R
\name{flyspin}
\alias{flyspin}
\title{Fly body turn calculator}
\usage{
flyspin(trx, framelag = 1)
}
\arguments{
\item{trx}{a track data frame (only requires CTrax output variables)}

\item{framelag}{the number of frames to compare angles across; defaults to 1 (consecutive frames)}
}
\value{
A numeric vector matching the structure of \code{trx} containing the spin value for each object and timepoint, or \code{NA}s when it can't be calculated.
}
\description{
Calculates the change in body angle between consecutive frames.
}
\details{
Uses the major axis angle of the ellipse fit by CTrax (\code{theta}) as this
turns out to be much more reliable than the angular speed approach based on centroid
location, especially when there is little movement.

Assumes that flies don't turn more than 90 degrees, so 'flips' the fly if that results in a lower angle.
Note that this assumption is probably true on a frame-by-frame basis; therefore if \code{framelag} is set
to a value >1, it is calculated as the sum of consecutive (small) spins.

It is currently run across the whole dataset without splitting it by \code{id}, as the amount of artefacts should
be small, but ideally it would be best to run it with an \code{ave}-based design.
}

