% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyturn.R
\name{flyturn}
\alias{flyturn}
\title{Track angular speed calculator}
\usage{
flyturn(trx, framelag = 6)
}
\arguments{
\item{trx}{a track data frame (only requires CTrax output variables)}

\item{framelag}{the number of frames to derive across; defaults to 6}
}
\value{
A numeric vector matching the structure of \code{trx} containing the angular speed value for each object and timepoint, or \code{NA}s when it can't be calculated.
}
\description{
Calculates the angular speed in rad/frame by comparing the angle of velocity vectors calculated across \code{framelag}.
}
\details{
For each timepoint t the speed vector is calculated between \code{t-framelag} and \code{t}, and then between \code{t} and \code{t+framelag}.
The angles are extracted from these two vectors and subtracted to yield the change in angle over \code{timeframe}.
This creates a lot of problems when the angles are close to the discontinuity of 2pi and needs sorting out.
Currently this leaves \code{framelag NA}s at the beginning and end of each track. This causes all sorts of problems
and may be best addressed somehow in the future.

Realistic data observation suggests that any angles over pi/2 are either artefacts or discontinuity errors, but at the moment they are not \code{NA}'d.
}

