% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narunmed.R
\name{narunmed}
\alias{narunmed}
\title{NA-tolerant running median}
\usage{
narunmed(xin, w)
}
\arguments{
\item{xin}{a numeric vector.}

\item{w}{the (odd numbered) median window to be passed to \code{runmed}.}
}
\value{
A vector with the same size and structure as the original (reconstructed by plonking NAs appropriately)
but running median-filtered values.
}
\description{
A serialising function that applies the \code{\link{runmed}} function to each consecutive stretch of
non-NA values in a numeric vector.
}

