% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIquadrant.R
\name{PIquadrant}
\alias{PIquadrant}
\title{Preference Index calculator}
\usage{
PIquadrant(trx, by.fly = F)
}
\arguments{
\item{trx}{input track data frame (must include \code{preference}, \code{sectiontimer} and \code{rept}).
Realistically it also requires the usual \code{trx} columns (in particular \code{id})}

\item{by.fly}{boolean to decide whether you calculate PI by section (of the 4 in the standard protocol,
arranged by section id x repeat)
or by fly id (collating all sections together for each fly)}
}
\value{
A data frame with the PI for each fly (if \code{by.fly} is true) or by section and repeat
(if \code{by.fly} is false).
}
\description{
Calculates preference indices for alternating quadrant sections of optogenetic protocol (section 4 and 5).
}
\details{
Frankly I can't remember what happens in this function and that is why you should document
this shit when you write it. I think it picks out the last 5s of each quadrant section and averages the distribution for each fly
(\code{by.fly=T}) or each repeat of the protocol (\code{by.fly=F}). However I see a lot of weird \code{NA}-based analysis in the code
and I can't remember what it's for.
}

