% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safemoves.R
\name{safemoves}
\alias{safemoves}
\title{Track motion parameter calculator}
\usage{
safemoves(trx, speedlag = 6, anglelag = 6, turnlag = 6)
}
\arguments{
\item{trx}{a track data frame (CTrax values)}

\item{speedlag}{frame lag for the calculation of speed}

\item{anglelag}{frame lag for the calculation of the velocity vector angle}

\item{turnlag}{frame lag for the calculation of the angular velocity}
}
\value{
A track data frame including the columns \code{speed}, \code{angle} and \code{turn}. Every \code{NA} entry will be removed.
The overwhelming majority of these will be \code{framelag} chunks at the beginning and end of tracks (where speed can't be calculated
because of lack of hindsight information) which don't affect the data structure much, but some would be artefactual entries culled by
the \code{speed>50} or \code{radius<400} sanity check. These could leave holes in the middle of tracks whereby consecutive rows would not be consecutive frames.
}
\description{
A wrapper function that calls \code{\link{flyrun}}, \code{\link{flyangle}} and \code{\link{flyturn}} and appends them to a track data frame, then removes any entry containing even one \code{NA}. Used for follow-up analysis that can't handle NAs.
}
\details{
Each motion function can receive a specific \code{framelag} value; by default they are all at 6.
This function is kept separate from \code{\link{trxgarnish}} so that it can be run with different \code{framelag} parameters without recalculating all the location variables as well.
}

