% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timewriter.R
\name{timewriter}
\alias{timewriter}
\title{Time protocol writer}
\usage{
timewriter(lit, prot, sec3 = 0)
}
\arguments{
\item{lit}{a boolean or 0-1 integer vector indicating frames with lights on or off}

\item{prot}{a list of protocol steps, in order; it can contain 0's and 1's, but only 2-5 are actually used}

\item{sec3}{an optional section 3 add-on to stick the last 2s back onto the protocol (input in frames!!!)}
}
\value{
A data frame containing the columns \code{time} (frame number), \code{lit} (lights on or off, 0-1 integer),
\code{chunk} (broad section definer including on frames and surrounding off frames), \code{section} (narrow section
definer attributing values to each on and off time separately), \code{rept} (repeat: used in conjunction with \code{section}
to uniquely identify each lights-on event) and various timers centered at the appropriate light-on time (\code{chunktimer},
\code{sectiontimer}, and \code{lighttimer}).
}
\description{
Creates and collates time protocol data given minimal information (light pattern and protocol steps)
}
\details{
Protocol steps are assumed to follow the scheme:

\itemize{

\item{0}{: no video (disregarded)}
\item{1}{: no light (disregarded)}
\item{2}{: lights constantly on}
\item{3}{: lights on in a blinking pattern}
\item{4}{: quadrants on, NE and SW}
\item{5}{: quadrants on, NW and SE}
}

This is primarily for the collapse of 4 and 5 into a single chunk.

There is currently no data cleanup. The underlying noise therefore shows up in frame variability (~4 frame jitter).
}

