% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trxgarnish.R
\name{trxgarnish}
\alias{trxgarnish}
\title{Track garnisher for time and position measures}
\usage{
trxgarnish(explist, t = NULL, filter = TRUE)
}
\arguments{
\item{explist}{a track data frame or experiment list (containing a \code{trx} data frame element)}

\item{t}{an optional time protocol data frame. It must have a \code{time} and an \code{id} column
which must match equivalent columns in the track data frame.}
}
\value{
A garnished track data frame or experiment list (depending on the format in input) including
the columns \code{radius}, \code{edge}, \code{distance} and \code{quadrant}, and the first and second
derivative variables \code{step}, \code{sidestep} and \code{spin}; if there is a time protocol, also
includes all columns of that data frame and the \code{preference} column calculated for the appropriate protocol section.
}
\description{
Computes distance from centre, quadrant edge and quadrant identity. Frame-by-frame variables are computed
(frame-by frame speed (\code{step}), angle relative to plate (\code{sidestep}) and change in theta (\code{spin})).
and also used for some quality control (eliminate speed>50). For movement calculated over different lags, see
\code{\link{flymoves}}).
}
\details{
If a time protocol is available, it also includes protocol structure information and calculates quadrant preference
(assuming quandrant patterns in sections 4 and 5 only).

Input can be either a \code{trx} data frame or an experiment list containing a \code{trx} data frame
and an \code{exp} list of experimental parameters. In the latter case some info is added
to the list, namely the arena parameters (in \code{$exp$arena}) and, if passed separately, time protocol (\code{$t}).

A time protocol can be passed to the function as an optional var or included in the original list
as an element named \code{$t}. It is merged into the track data frame by \code{id} and \code{time}.
If the time protocol contains a \code{section} column, this is combined with quadrant information to
compute \code{preference} for lit quadrant (assumed to be positive diagonal for section 4 and
negative diagonal for section 5).
}

