% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trxread.R
\name{trxread}
\alias{trxread}
\title{CTrax input utility}
\usage{
trxread(trxfile = file.choose())
}
\arguments{
\item{trxfile}{filename: defaults to browsing window}
}
\value{
A list object containing a list of experimental details (\code{$exp}) and a data frame
of tracking data (\code{$trx}). \code{exp} contains the filename and, for \code{fixerrors}-processed
data, video filename, framerate, pixel size. \code{trx} always contains the same format of
tracking data with the columns \code{id}, \code{x}, \code{y}, \code{a}, \code{b}, \code{theta} and \code{time}.
}
\description{
Input function for \code{.mat} files generated by CTrax.
Creates a list object including experimental parameters and tracking data.
}
\details{
Uses \code{readMat} from \code{\link{R.matlab}} to read the file.

CTrax files exist in two very different formats depending on whether they have been processed
through the \code{fixerrors} GUI or not. This input function can read both but some data will
only be available for the latter (video info, framerate etc).
}

